reorganize integration tests
mockwebserver gets its own package
only run macos tests in integration tests so that we have access to the
test fixtures
