+++ .idea/runConfigurations/MPP_Integration_Tests.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="MPP Integration Tests" type="GradleRunConfiguration" factoryName="Gradle">
     <ExternalSystemSettings>
       <option name="executionName" />
-      <option name="externalProjectPath" value="$PROJECT_DIR$/apollo-integration-kotlin" />
+      <option name="externalProjectPath" value="$PROJECT_DIR$/integration-tests-kotlin" />
       <option name="externalSystemIdString" value="GRADLE" />
       <option name="scriptParameters" value="" />
       <option name="taskDescriptions">
