+++ composite/integration-tests-kotlin/src/appleTest/kotlin/com/apollographql/apollo3/integration/Cwd.kt
@@ -1,0 +1,17 @@
+package com.apollographql.apollo3.integration
+
+import kotlinx.cinterop.ByteVar
+import kotlinx.cinterop.allocArray
+import kotlinx.cinterop.convert
+import kotlinx.cinterop.memScoped
+import kotlinx.cinterop.toKString
+import platform.posix.getcwd
+
+
+fun cwd(): String {
+  return memScoped {
+    val maxSize = 1024
+    val buf = allocArray<ByteVar>(maxSize)
+    getcwd(buf, maxSize.convert())?.toKString() ?: "?"
+  }
+}
