+++ composite/integration-tests-kotlin/src/commonTest/kotlin/com/apollographql/apollo3/integration/mockserver/MockServer.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.integration
+package com.apollographql.apollo3.integration.mockserver
 
 expect class MockServer() {
   fun url(): String
