+++ composite/integration-tests-kotlin/src/commonTest/kotlin/com/apollographql/apollo3/integration/test/FetchPolicyTest.kt
@@ -7,7 +7,7 @@ import com.apollographql.apollo3.cache.normalized.ApolloStore
 import com.apollographql.apollo3.cache.normalized.CacheKeyResolver
 import com.apollographql.apollo3.cache.normalized.MemoryCacheFactory
 import com.apollographql.apollo3.cache.normalized.internal.ApolloStore
-import com.apollographql.apollo3.integration.MockServer
+import com.apollographql.apollo3.integration.mockserver.MockServer
 import com.apollographql.apollo3.integration.enqueue
 import com.apollographql.apollo3.interceptor.cache.FetchPolicy
 import com.apollographql.apollo3.interceptor.cache.isFromCache
